Refactor form functionality into domManipulation
