+++ src/scripts.js
@@ -6,119 +6,13 @@ import './images/Sleep.svg';
 import './images/Friends.svg';
 import './images/activity.svg'
 import './images/Plus.svg';
-import {displayProfileInfo, displayStepInfo, displayHydrationInfo, displaySleepInfo, displayActivityInfo, displayFriendsInfo, scroll } from './domManipulation';
+import {displayProfileInfo, displayStepInfo, displayHydrationInfo, displaySleepInfo, displayActivityInfo, displayFriendsInfo, scroll, formListen } from './domManipulation';
 import { fetchData, postData } from './fetch';
 import DataManager from './DataManager';
 import UserRepository from './UserRepository';
 import { Chart, registerables } from 'chart.js';
 Chart.register(...registerables);
 
-// Query Selectors
-const userProfile = document.querySelector('#userProfile');
-const activityTypeForm = document.querySelector('#activity-type-form');
-const activityOptions = document.querySelector('#activity-type');
-const sleepForm = document.querySelector('#sleep-form');
-const hydrationForm = document.querySelector('#hydration-form');
-const activityForm = document.querySelector('#activity-form');
-const userInputHoursSlept = document.querySelector('#hours-slept');
-const userInputSleepDate = document.querySelector('#sleep-date');
-const userInputSleepQuality = document.querySelector('#sleep-quality');
-const userInputHydrationDate = document.querySelector('#hydration-date');
-const userInputHydrationOunces = document.querySelector('#num-ounces');
-const userInputActivityDate = document.querySelector('#activity-date');
-const userInputActivityStairs = document.querySelector('#flight-of-stairs');
-const userInputActivityMinutesActive = document.querySelector('#minutes-active');
-const userInputActivitySteps = document.querySelector('#number-of-steps');
-const confirmationText = document.querySelector('#confirmation-text');
-const errorText = document.querySelector('#error-text');
-
-const formatDate = (date) => {
-  const formattedDate = date.replaceAll('-', '/');
-  return formattedDate;
-}
-
-const toggleConfirmationText = (result) => {
-  result.hidden = !result.hidden;
-}
-
-const checkFetchResult = (result) => {
-  if (result === 'SUCCESS') {
-    toggleConfirmationText(confirmationText);
-    setTimeout(() => {
-      toggleConfirmationText(confirmationText);
-    }, 4000);
-  } else {
-    toggleConfirmationText(errorText);
-    setTimeout(() => {
-      toggleConfirmationText(errorText);
-    }, 4000);
-  }
-}
-
-activityTypeForm.addEventListener('change', () => {
-  switch (activityOptions.value) {
-    case 'sleep':
-      hydrationForm.hidden = true;
-      activityForm.hidden = true;
-      sleepForm.hidden = false;
-      break;
-
-      case 'hydration':
-        sleepForm.hidden = true;
-        activityForm.hidden = true;
-        hydrationForm.hidden = false;
-        break;
-
-      case 'activity':
-        sleepForm.hidden = true;
-        hydrationForm.hidden = true;
-        activityForm.hidden = false;
-      break;
-
-    default:
-      console.log('something went wrong');
-      break;
-  }
-});
-
-sleepForm.addEventListener('submit', (event) => {
-  event.preventDefault();
-  const userInputSleepData = {
-    userID: userRepo.activeUser.id,
-    date: formatDate(userInputSleepDate.value),
-    hoursSlept: userInputHoursSlept.value,
-    sleepQuality: userInputSleepQuality.value
-  }
-  sleepForm.reset();
-  postData('sleep', userInputSleepData)
-    .then(checkFetchResult);
-});
-
-hydrationForm.addEventListener('submit', (event) => {
-  event.preventDefault();
-  const userInputHydrationData = {
-    userID: userRepo.activeUser.id,
-    date: formatDate(userInputHydrationDate.value),
-    numOunces: userInputHydrationOunces.value
-  }
-  hydrationForm.reset();
-  postData('hydration', userInputHydrationData)
-    .then(checkFetchResult);
-});
-
-activityForm.addEventListener('submit', (event) => {
-  event.preventDefault();
-  const userInputActivityData = {
-    userID: userRepo.activeUser.id,
-    date: formatDate(userInputActivityDate.value),
-    flightsOfStairs: userInputActivityStairs.value,
-    minutesActive: userInputActivityMinutesActive.value,
-    numSteps: userInputActivitySteps.value,
-  }
-  activityForm.reset();
-  postData('activity', userInputActivityData)
-    .then(checkFetchResult);
-});
 
 // Global Variables
 const userRepo = new UserRepository();
@@ -151,10 +45,11 @@ const renderDOM = (dataManager) => {
 
 const displayAllUserInfo = (user, userRepo) => {
   displayProfileInfo(user);
-  displayActivityInfo(user, userRepo, dataManager)
+  displayActivityInfo(user, userRepo, dataManager);
   displayStepInfo(user, userRepo);
   displayHydrationInfo(user);
   displaySleepInfo(user);
+  formListen(userRepo);
 }
 
 const getRandomIndex = (array) => {
